// src/style/types.ts
export type MarkerPlacement = 'interval' | 'vertex' | 'innerVertex' | 'first' | 'last' | 'center' | 'segmentCenter';